requirements.txt
# =============================================
# Librerías Base
# =============================================
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
idna==3.10
itsdangerous==2.2.0
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
typing-inspection==0.4.0
typing_extensions==4.13.1
schedule==1.2.1

# =============================================
# Motor Principal AXXON (IA Simbólica)
# =============================================
openai==1.30.5
pydantic==2.11.2
pydantic_core==2.33.1

# =============================================
# Frameworks Web (APIs y Server)
# =============================================
Flask==3.1.0
gunicorn==21.2.0
streamlit==1.35.0

# =============================================
# Procesamiento de Lenguaje Natural y Vectores
# =============================================
transformers==4.40.1
faiss-cpu==1.7.4
numpy==1.26.4
tqdm==4.66.4

# =============================================
# Clientes API y Bases de Datos
# =============================================
notion-client==2.3.0
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pandas==2.2.2
gspread==5.12.0
oauth2client==4.1.3
pytz==2024.1

# =============================================
# Documentos y PDF
# =============================================
PyMuPDF==1.24.2
fpdf==1.7.2
PyPDF2==3.0.1
python-docx==1.1.2

# =============================================
# Integraciones Externas
# =============================================
httpcore==1.0.7
httpx==0.28.1
jiter==0.9.0
serpapi==0.1.5

# =============================================
# Web Scraping y Word Documents
# =============================================
fitz==0.0.1.dev2